+++ portal/templates/portal/teach/dashboard.html
@@ -194,9 +194,9 @@
                                     <small>
                                         {{ sent_invite.invited_teacher_first_name }}
                                         {% if sent_invite.is_expired %}
-                                        <b>(expired)</b>
+                                            <b>(expired)</b>
                                         {% else %}
-                                        <b>(pending)</b>
+                                            <b>(pending)</b>
                                         {% endif %}
                                     </small>
                                 </p>
