+++ portal/views/teacher/teach.py
@@ -923,100 +923,3 @@ def count_student_details_click(download_type):
         raise Exception("Unknown download type")
 
     activity_today.save()
-
-
-def process_teacher_invitation(request, token):
-    try:
-        invitation = SchoolTeacherInvitation.objects.get(token=token, expiry__gt=timezone.now())
-    except SchoolTeacherInvitation.DoesNotExist:
-        return "Uh oh, the Invitation does not exist or it has expired. ☹️"
-
-    if User.objects.filter(email=invitation.invited_teacher_email).exists():
-        return (
-            "It looks like an account is already registered with this email address. You will need to delete the "
-            "other account first or change the email associated with it in order to proceed. You will then be able to "
-            "access this page."
-        )
-    else:
-        if request.method == "POST":
-            invited_teacher_form = InvitedTeacherForm(request.POST)
-            if invited_teacher_form.is_valid():
-                data = invited_teacher_form.cleaned_data
-                invited_teacher_password = data["teacher_password"]
-                newsletter_ticked = data["newsletter_ticked"]
-
-                # Create the teacher
-                invited_teacher = Teacher.objects.factory(
-                    first_name=invitation.invited_teacher_first_name,
-                    last_name=invitation.invited_teacher_last_name,
-                    email=invitation.invited_teacher_email,
-                    password=invited_teacher_password,
-                )
-                invited_teacher.is_admin = invitation.invited_teacher_is_admin
-                invited_teacher.school = invitation.school
-                invited_teacher.invited_by = invitation.from_teacher
-                invited_teacher.save()
-
-                # Verify their email
-                generate_token(invited_teacher.new_user, preverified=True)
-
-                # Add to Dotmailer if they ticked the box
-                if newsletter_ticked:
-                    user = invited_teacher.user.user
-                    add_to_dotmailer(user.first_name, user.last_name, user.email, DotmailerUserType.TEACHER)
-
-                # Anonymise the invitation
-                invitation.anonymise()
-
-
-@login_required(login_url=reverse_lazy("teacher_login"))
-def delete_teacher_invite(request, token):
-    invite_to_delete = SchoolTeacherInvitation.objects.filter(token=token)[0]
-    invite_teacher_first_name = invite_to_delete.invited_teacher_first_name
-    invite_to_delete.anonymise()
-    messages.success(request, f"Invite for {invite_teacher_first_name} successfully deleted")
-    return HttpResponseRedirect(reverse_lazy("dashboard"))
-
-
-@login_required(login_url=reverse_lazy("teacher_login"))
-def resend_invite_teacher(request, token):
-    invitation = SchoolTeacherInvitation.objects.get(token=token)
-    invitation.expiry = timezone.now() + timedelta(days=30)
-    invitation.save()
-    teacher = Teacher.objects.filter(id=invitation.from_teacher.id)[0]
-    verify_email = EmailVerification(
-        user=teacher.new_user,
-        token=token,
-        email=invitation.invited_teacher_email,
-        expiry=timezone.now() + timedelta(hours=1),
-        verified=True,
-    )
-
-    messages.success(request, "Teacher re-invited!")
-    message = email_messages.inviteTeacherEmail(request, invitation.school, token, not (invitation.is_expired))
-    send_email(
-        INVITE_FROM,
-        [invitation.invited_teacher_email],
-        message["subject"],
-        message["message"],
-        message["subject"],
-    )
-    return HttpResponseRedirect(reverse_lazy("dashboard"))
-
-
-def invited_teacher(request, token):
-    error_message = process_teacher_invitation(request, token)
-
-    if request.method == "POST":
-        invited_teacher_form = InvitedTeacherForm(request.POST)
-        if invited_teacher_form.is_valid():
-            messages.success(request, "Your account has been created successfully, please log in.")
-            return HttpResponseRedirect(reverse_lazy("teacher_login"))
-    else:
-        invited_teacher_form = InvitedTeacherForm()
-
-    return render(
-        request,
-        "portal/teach/invited.html",
-        {"invited_teacher_form": invited_teacher_form, "error_message": error_message},
-    )
